Moved Room component out of App.js
